+++ tests/BasicAuthenticationTest.php
@@ -220,7 +220,7 @@ class HttpBasicAuthenticationTest extends \PHPUnit_Framework_TestCase
                 "root" => "t00r",
                 "user" => "passw0rd"
             ],
-            "after" => function ($request, $response, $arguments) {
+            "after" => function ($response, $arguments) {
                 return $response
                     ->withBody((new StreamFactory)->createStream())
                     ->withStatus(401)
@@ -255,7 +255,7 @@ class HttpBasicAuthenticationTest extends \PHPUnit_Framework_TestCase
                 "root" => "t00r",
                 "user" => "passw0rd"
             ],
-            "after" => function ($request, $response, $arguments) {
+            "after" => function ($response, $arguments) {
                 return $response->withHeader("X-Brawndo", "plants crave");
             }
         ]);
@@ -285,7 +285,7 @@ class HttpBasicAuthenticationTest extends \PHPUnit_Framework_TestCase
                 "root" => "t00r",
                 "user" => "passw0rd"
             ],
-            "error" => function ($request, $response, $arguments) {
+            "error" => function ($response, $arguments) {
                 $response->getBody()->write("ERROR: " . $arguments["message"]);
                 return $response;
             }
@@ -316,7 +316,7 @@ class HttpBasicAuthenticationTest extends \PHPUnit_Framework_TestCase
                 "root" => "t00r",
                 "user" => "passw0rd"
             ],
-            "error" => function ($request, $response, $arguments) {
+            "error" => function ($response, $arguments) {
                 return $response
                     ->withStatus(302)
                     ->withHeader("Location", "/foo/bar");
@@ -451,7 +451,7 @@ class HttpBasicAuthenticationTest extends \PHPUnit_Framework_TestCase
                 "root" => "t00r",
                 "user" => "passw0rd"
             ],
-            "before" => function ($request, $response, $arguments) {
+            "before" => function ($request, $arguments) {
                 return $request->withAttribute("user", $arguments["user"]);
             }
         ]);